refactoring msg-from-client
