+++ src/board_game_components/lobby_page.rs
@@ -8,12 +8,14 @@ use lib_rpg::server::server_manager::{GamePhase, ServerData};
 use crate::components::button::ButtonVariant;
 use crate::{
     board_game_components::{
-        character_select::CharacterSelect, common_comp::ButtonLink,
-        msg_from_client::send_start_game, startgame_page::StartGamePage,
+        character_select::CharacterSelect, common_comp::ButtonLink, startgame_page::StartGamePage,
     },
     common::{Route, SERVER_NAME},
     components::button::Button,
-    websocket_handler::event::{ClientEvent, ServerEvent},
+    websocket_handler::{
+        event::{ClientEvent, ServerEvent},
+        msg_from_client::send_start_game,
+    },
 };
 
 #[component]
