+++ src/websocket_handler/mod.rs
@@ -1,1 +1,2 @@
 pub mod event;
+pub mod msg_from_client;
